render_deploy_ready/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
astroid==2.15.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.0.1
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2022.12.7
chardet==5.2.0
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.6
dirtyjson==1.0.8
distro==1.9.0
Django==5.1.1
djangorestframework==3.15.2
etelemetry==0.3.1
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.12.2
fitz==0.0.1.dev2
Flask==3.0.0
fonttools==4.57.0
fpdf==1.7.2
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.30.2
idna==3.4
imbalanced-learn==0.13.0
imblearn==0.0
importlib_resources==6.4.5
ipykernel==6.26.0
ipython==8.17.2
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.9.0
joblib==1.4.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.8
lazy-object-proxy==1.9.0
llama-index-core==0.11.17
llama-index-readers-file==0.2.2
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==2.1.3
marshmallow==3.22.0
matplotlib==3.10.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
nibabel==5.3.0
nipype==1.8.6
nltk==3.9.1
numpy==1.23.5
openai==1.73.0
opencv-python==4.7.0.72
packaging==23.2
pandas==2.2.3
parso==0.8.3
pathlib==1.0.1
Pillow==9.3.0
platformdirs==3.5.0
prompt-toolkit==3.0.41
propcache==0.2.0
prov==2.0.1
psutil==5.9.6
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.10.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
Pygments==2.16.1
pylint==2.17.3
PyMuPDF==1.24.11
pyparsing==3.1.4
pypdf==4.3.1
pytesseract==0.3.13
python-dateutil==2.8.2
pytz==2022.7.1
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==25.1.1
rdflib==6.3.2
regex==2024.9.11
reportlab==4.2.5
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
simplejson==3.19.3
six==1.16.0
sketchpy==0.1.0
sklearn-compat==0.1.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.39.2
striprtf==0.0.26
svg.path==6.2
svgpathtools==1.5.1
svgwrite==1.4.3
sympy==1.12
tenacity==8.5.0
tesseract==0.1.3
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.11.8
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.5
traitlets==5.13.0
traits==6.3.2
transformers==4.30.2
turtle==0.0.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.15
uvicorn==0.31.1
wcwidth==0.2.10
Werkzeug==3.0.1
wrapt==1.15.0
xgboost==3.0.0
yarl==1.14.0